requirements.txt
apturl==0.5.2
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
attrs==22.1.0
autopep8==2.0.1
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.3
django-auth-ldap==4.1.0
django-bootstrap-v5==1.0.11
django-tinymce==3.5.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.8.0
future==0.18.2
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
oscrypto==1.3.0
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.2
pluggy==0.13.0
protobuf==3.12.4
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
pyHanko==0.17.0
pyhanko-certvalidator==0.20.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==3.3.0
pypng==0.20220715.0
pyRFC3339==1.1
pyrsistent==0.19.2
python-apt==2.4.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-ldap==3.3.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.4
reportlab==3.6.12
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
svglib==1.5.1
systemd-python==234
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tox==3.21.4
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
# ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritools==4.0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.16.5
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xhtml2pdf==0.2.9
xkit==0.0.0
zipp==1.0.0
